backend/requirement.txt
fastapi
uvicorn
httpx
openai
python-dotenv
pydantic